+++ TravelMeitApp/Colors.swift
@@ -1,0 +1,15 @@
+//
+//  Colors.swift
+//  TravelMeitApp
+//
+//  Created by MAC1DIGITAL20 on 28/05/23.
+//
+
+import Foundation
+import SwiftUI
+
+extension Color {
+    static let principalColor = Color("PrincipalColor")
+    static let secondColor = Color("SecondColor")
+    
+}
