--- TravelMeitApp/ContentView.swift
@@ -1,26 +1,0 @@
-//
-//  ContentView.swift
-//  TravelMeitApp
-//
-//  Created by MAC1DIGITAL20 on 17/05/23.
-//
-
-import SwiftUI
-
-struct ContentView: View {
-    var body: some View {
-        VStack {
-            Image(systemName: "globe")
-                .imageScale(.large)
-                .foregroundColor(.accentColor)
-            Text("Hello, world!")
-        }
-        .padding()
-    }
-}
-
-struct ContentView_Previews: PreviewProvider {
-    static var previews: some View {
-        ContentView()
-    }
-}
